Use the undocumented `--test-app-path` option
Since ffb6fe9, the jade compiler is split in two packages: `jade`
and `jade-compiler`. This is useful for third-party packages that
want to use the compiler directly without the source handler (eg
`meteorhacks:ssr`). However it makes testing more complicated. The
problem is that `mquandalle:jade` use `mquandalle:jade-compiler` as
a dependency, so if I go to the `jade` package and test it using the
classical `meteor test-packages ./`, meteor will download the
`mquandalle:jade-compiler` from atmosphere instead of using the local
version. Since these two packages are highly integrated, a change in
the compiler will certainly have consequences in the source handler
and I don't want to release a - potentially broken - test version of
the compiler every time I want to test it with the source handler.

This commit solves the above problem thanks to the undocumented
`--test-app-path` option of the `test-packages` command. This option
allow us to test several local packages simultaneously without having
to release+publish+require them. We put all the packages we want to
test in a `packages/` directory — which is by the way the organization
used by meteor-core. We then use the following command:

  meteor test-packages --test-app-path . packages/*

The `--test-app-path .` option let create the tinytest application in
the root of the repository. Because of this when a package `A` require
a package `B`, meteor will first look in the `packages/` directory as
it does for normal meteor applications. We then test all the local
packages using the star selector: `packages/*`.

This commit also add the whole `.meteor` directory in the `.gitignore`
list because we don't care of the test application (by default this
application is created in a temporary directory outside the
repository).