Fix file caching
One of the main advantage of the new build plugin API introduced in
Meteor 1.2 was the possibility to define "incremental compilers" that
relies on caching to accelerate the build process.

Unfortunately 7734099 didn't implement this new caching mechanism
correctly, this commit should make it work.

The documentation for CachingCompiler is available at:

  https://github.com/meteor/meteor/tree/devel/packages/caching-compiler